feat: fix connect nodes functionality

+++ src/main/java/com/example/virtu360/tour/interfaces/rest/transform/LinkResourceFromEntityAssembler.java
@@ -12,8 +12,8 @@ public class LinkResourceFromEntityAssembler {
     return new LinkResource(
         entity.getFromNode().getId().toString(),
         entity.getToNodeId(),
-        (int)entity.getPosition().yaw(),
-        (int)entity.getPosition().pitch()
+        entity.getPosition().yaw(),
+        entity.getPosition().pitch()
     );
   }
 }
